LayoutTests/inspector/styles/media-queries-expected.txt

Tests that media query stack is rendered for associated rules.

Main style:
[expanded] 
element.style  { ()

======== Matched CSS Rules ========
[expanded] 
media="all" (media-queries.html)
@media screen (media-queries.css:5)
#main  { (media-queries.css:6)
color: yellow;

[expanded] 
media="all" (media-queries.html)
@import screen (media-queries.css:3)
@media not aural (media-queries-1.css:5)
#main  { (media-queries-1.css:6)
border: 1px solid black;

[expanded] 
media="screen" (media-queries.html)
@media not print (media-queries.html:5)
#main  { (media-queries.html:6)
background: blue;

[expanded] 
div  { (user agent stylesheet)
display: block;
